products fragment and logic is add

+++ app/src/main/java/com/example/myapplication/dataBase/Equipment.kt
@@ -14,7 +14,7 @@ data class Equipment(
     @ColumnInfo(name = "photo")
     val photo : String,
     @ColumnInfo(name = "weight")
-    val weight : Double,
+    val weight : Int,
     @ColumnInfo(name = "equipment_in_the_campaign")
     val equipmentInTheCampaign : Boolean
