+++ app/src/main/java/com/example/myapplication/ui/adding_a_participant/AddingAParticipantViewModel.kt
@@ -31,8 +31,8 @@ class AddingAParticipantViewModel(private val hikeDao: HikeDao) : ViewModel() {
         lastName: String,
         gender: String,
         age: String,
-        maximumPortableWeight: Double,
-        weightOfPersonalItems: Double,
+        maximumPortableWeight: Int,
+        weightOfPersonalItems: Int,
         participationInTheCampaign: Boolean
     ) {
         ParticipantsEquipmentUseCase(hikeDao).loadParticipants(
