+++ app/src/main/java/com/example/myapplication/ui/adding_equipment/AddingEquipmentViewModel.kt
@@ -14,9 +14,6 @@ import kotlinx.coroutines.launch
 
 class AddingEquipmentViewModel(private val hikeDao: HikeDao) : ViewModel() {
 
-//    suspend fun getAllEquipmentFlow(): Flow<List<Equipment>> {
-//        return ParticipantsEquipmentUseCase(hikeDao).getAllCollectionEquipmentFlow()
-//    }
 
     suspend fun getAllEquipmentList(): List<Equipment> {
         return ParticipantsEquipmentUseCase(hikeDao).getAllCollectionEquipmentList()
@@ -29,7 +26,7 @@ class AddingEquipmentViewModel(private val hikeDao: HikeDao) : ViewModel() {
         id: Int,
         name: String,
         photo: String,
-        weight: Double,
+        weight: Int,
         equipmentInTheCampaign: Boolean
 
     ) {
